+++ CHANGELOG.md
@@ -7,7 +7,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 ## [Unreleased]
 ### Added
-- `ignoreSpecificationFilters` on `SearchQuery`.
+- `facetsBehavior` on `SearchQuery`.
 
 ## [3.40.0] - 2019-12-27
 ### Added